+++ lib/src/classes/request.dart
@@ -1,15 +1,15 @@
 import 'serializable.dart';
 import 'response.dart';
 
-abstract class OBSWebSocketRequest<T extends OBSWebSocketResponse>
+abstract class ObsWebSocketRequest<T extends ObsWebSocketResponse>
     extends Serializable {
   final String type;
 
   T? response;
 
-  OBSWebSocketRequest(this.type, [Map<String, dynamic>? data])
+  ObsWebSocketRequest(this.type, [Map<String, dynamic>? data])
       : super(data ?? {});
 
-  OBSWebSocketResponse serializeResponse(data, status) =>
-      OBSWebSocketResponse(data, status);
+  ObsWebSocketResponse serializeResponse(data, status) =>
+      ObsWebSocketResponse(data, status);
 }