+++ lib/src/classes/response.dart
@@ -1,8 +1,8 @@
 import 'serializable.dart';
 import '../ops/request_response.dart';
 
-class OBSWebSocketResponse extends Serializable {
-  final RequestResponseStatus status;
+class ObsWebSocketResponse extends Serializable {
+  final ObsRequestResponseStatus status;
 
-  OBSWebSocketResponse(super.data, this.status);
+  ObsWebSocketResponse(super.data, this.status);
 }