+++ lib/src/ops/event.dart
@@ -3,13 +3,13 @@ import '../protocol/enums.dart';
 import 'ops.dart';
 
 /// An event coming from OBS has occured. Eg scene switched, source muted.
-class EventOpCode extends OpCode {
+class ObsEventOp extends ObsOp {
   String get eventType => data["eventType"];
   int get eventIntent => data["eventIntent"];
   Map<String, dynamic>? get eventData => data["eventData"];
 
   @override
-  WebSocketOpCode get code => WebSocketOpCode.event;
+  ObsWebSocketOpCode get code => ObsWebSocketOpCode.event;
 
-  EventOpCode(super.data);
+  ObsEventOp(super.data);
 }