+++ lib/src/ops/hello.dart
@@ -5,7 +5,7 @@ import 'ops.dart';
 /// First message sent from the server immediately on client connection.
 /// Contains authentication information if auth is required.
 /// Also contains RPC version for version negotiation.
-class HelloOpCode extends OpCode {
+class ObsHelloOp extends ObsOp {
   /// obs-websocket version of the host
   String get obsWebSocketVersion => data["obsWebSocketVersion"];
 
@@ -18,9 +18,9 @@ class HelloOpCode extends OpCode {
       : null;
 
   @override
-  WebSocketOpCode get code => WebSocketOpCode.hello;
+  ObsWebSocketOpCode get code => ObsWebSocketOpCode.hello;
 
-  HelloOpCode(super.data);
+  ObsHelloOp(super.data);
 }
 
 /// Authentication info sent by HelloOp required to the authenticate the client