+++ lib/src/ops/ops.dart
@@ -2,8 +2,8 @@ import '../obs_websocket.dart';
 import '../classes/serializable.dart';
 
 export 'hello.dart';
-export 'identified.dart';
 export 'identify.dart';
+export 'identified.dart';
 export 'reidentify.dart';
 export 'event.dart';
 export 'request.dart';
@@ -11,21 +11,21 @@ export 'request_response.dart';
 export 'request_batch.dart';
 export 'request_batch_response.dart';
 
-abstract class OpCode extends Serializable {
-  WebSocketOpCode get code;
+abstract class ObsOp extends Serializable {
+  ObsWebSocketOpCode get code;
 
-  OpCode(super.data);
+  ObsOp(super.data);
 
-  static const Map<WebSocketOpCode, OpCode Function(Map<String, dynamic> data)>
-      opCodeMap = {
-    WebSocketOpCode.hello: HelloOpCode.new,
-    WebSocketOpCode.identify: IdentifyOpCode.new,
-    WebSocketOpCode.identified: IdentifiedOpCode.new,
-    WebSocketOpCode.reidentify: ReidentifyOpCode.new,
-    WebSocketOpCode.event: EventOpCode.new,
-    WebSocketOpCode.request: RequestOpCode.new,
-    WebSocketOpCode.requestResponse: RequestResponseOpCode.new,
-    WebSocketOpCode.requestBatch: RequestBatchOpCode.new,
-    WebSocketOpCode.requestBatchResponse: RequestBatchResponseOpCode.new
+  static const Map<ObsWebSocketOpCode,
+      ObsOp Function(Map<String, dynamic> data)> opCodeMap = {
+    ObsWebSocketOpCode.hello: ObsHelloOp.new,
+    ObsWebSocketOpCode.identify: ObsIdentifyOp.new,
+    ObsWebSocketOpCode.identified: ObsIdentifiedOp.new,
+    ObsWebSocketOpCode.reidentify: ObsReidentifyOp.new,
+    ObsWebSocketOpCode.event: ObsEventOp.new,
+    ObsWebSocketOpCode.request: ObsRequestOp.new,
+    ObsWebSocketOpCode.requestResponse: ObsRequestResponseOp.new,
+    ObsWebSocketOpCode.requestBatch: ObsRequestBatchOp.new,
+    ObsWebSocketOpCode.requestBatchResponse: ObsRequestBatchResponseOp.new
   };
 }