+++ lib/src/ops/reidentify.dart
@@ -3,13 +3,13 @@ import 'ops.dart';
 
 /// Sent at any time after initial identification to update the provided session
 /// parameters.
-class ReidentifyOpCode extends OpCode {
+class ObsReidentifyOp extends ObsOp {
   int get eventSubscriptions => data["eventSubscriptions"];
 
   @override
-  WebSocketOpCode get code => WebSocketOpCode.reidentify;
+  ObsWebSocketOpCode get code => ObsWebSocketOpCode.reidentify;
 
-  ReidentifyOpCode(super.data);
-  ReidentifyOpCode.create(int eventSubscriptions)
+  ObsReidentifyOp(super.data);
+  ObsReidentifyOp.create(int eventSubscriptions)
       : super({"eventSubscriptions": eventSubscriptions});
 }