+++ lib/src/ops/request.dart
@@ -3,16 +3,16 @@ import 'ops.dart';
 
 /// Client is making a request to obs-websocket. Eg get current scene, create
 /// source.
-class RequestOpCode extends OpCode {
+class ObsRequestOp extends ObsOp {
   String get requestType => data["requestType"];
   String get requestId => data["requestId"];
   Map<String, dynamic>? get requestData => data["requestData"];
 
   @override
-  WebSocketOpCode get code => WebSocketOpCode.request;
+  ObsWebSocketOpCode get code => ObsWebSocketOpCode.request;
 
-  RequestOpCode(super.data);
-  RequestOpCode.create(
+  ObsRequestOp(super.data);
+  ObsRequestOp.create(
       {required String requestType,
       required String requestId,
       Map<String, dynamic>? requestData})