+++ lib/src/ops/request_batch.dart
@@ -3,25 +3,25 @@ import 'ops.dart';
 
 /// Client is making a batch of requests for obs-websocket.
 /// Requests are processed serially (in order) by the server.
-class RequestBatchOpCode extends OpCode {
+class ObsRequestBatchOp extends ObsOp {
   String get requestId => data["requestId"];
   bool get haltOnFailure => data["haltOnFailure"] ?? false;
-  RequestBatchExecutionType get executionType =>
-      RequestBatchExecutionType.fromInt(data["executionType"] ??
-          RequestBatchExecutionType.serialRealtime.value);
-  List<RequestOpCode> get requests =>
-      data["requests"].map((d) => RequestOpCode(d)).toList();
+  ObsRequestBatchExecutionType get executionType =>
+      ObsRequestBatchExecutionType.fromInt(data["executionType"] ??
+          ObsRequestBatchExecutionType.serialRealtime.value);
+  List<ObsRequestOp> get requests =>
+      data["requests"].map((d) => ObsRequestOp(d)).toList();
 
   @override
-  WebSocketOpCode get code => WebSocketOpCode.requestBatch;
+  ObsWebSocketOpCode get code => ObsWebSocketOpCode.requestBatch;
 
-  RequestBatchOpCode(super.data);
-  RequestBatchOpCode.create(
+  ObsRequestBatchOp(super.data);
+  ObsRequestBatchOp.create(
       {required String requestId,
       bool haltOnFailure = false,
-      RequestBatchExecutionType executionType =
-          RequestBatchExecutionType.serialRealtime,
-      required List<RequestOpCode> requests})
+      ObsRequestBatchExecutionType executionType =
+          ObsRequestBatchExecutionType.serialRealtime,
+      required List<ObsRequestOp> requests})
       : super({
           "requestId": requestId,
           "haltOnFailure": haltOnFailure,